refactor(web): extract submappers to dir
