+++ appeals/web/src/server/lib/mappers/__tests__/mappers.test.js
@@ -5,7 +5,7 @@ import {
 } from '#testing/app/fixtures/referencedata.js';
 import { createAccountInfo } from '#testing/app/app.js';
 import { initialiseAndMapAppealData } from '../appeal/appeal.mapper.js';
-import { initialiseAndMapLPAQData } from '../lpaQuestionnaire.mapper.js';
+import { initialiseAndMapLPAQData } from '../lpa-questionnaire/lpa-questionnaire.mapper.js';
 import { areIdsDefinedAndUnique } from '#testing/lib/testMappers.js';
 import {
 	buildNotificationBanners,
