+++ appeals/web/src/server/lib/mappers/lpa-questionnaire/submappers/map-conservation-area-map.js
@@ -1,0 +1,12 @@
+import { documentInstruction } from '../utils.js';
+
+/** @type {import("../lpa-questionnaire.mapper.js").SubMapper} */
+export const mapConservationAreaMap = ({ lpaQuestionnaireData, session }) =>
+	documentInstruction({
+		id: 'conservation-area-map',
+		text: 'Conservation area map and guidance',
+		folderInfo: lpaQuestionnaireData.documents.conservationMap,
+		cypressDataName: 'conservation-area-map-and-guidance',
+		lpaQuestionnaireData,
+		session
+	});
