+++ appeals/web/src/server/lib/mappers/lpa-questionnaire/submappers/map-letters-to-neighbours.js
@@ -1,0 +1,12 @@
+import { documentInstruction } from '../utils.js';
+
+/** @type {import("../lpa-questionnaire.mapper.js").SubMapper} */
+export const mapLettersToNeighbours = ({ lpaQuestionnaireData, session }) =>
+	documentInstruction({
+		id: 'letters-to-neighbours',
+		text: 'Letter or email notification',
+		folderInfo: lpaQuestionnaireData.documents.whoNotifiedLetterToNeighbours,
+		cypressDataName: 'letter-or-email-notification',
+		lpaQuestionnaireData,
+		session
+	});
