+++ appeals/web/src/server/lib/mappers/lpa-questionnaire/submappers/map-lpa-health-and-safety.js
@@ -1,0 +1,15 @@
+import { booleanWithDetailsSummaryListItem } from '#lib/mappers/components/boolean.js';
+
+/** @type {import("../lpa-questionnaire.mapper.js").SubMapper} */
+export const mapLpaHealthAndSafety = ({ lpaQuestionnaireData, currentRoute, userHasUpdateCase }) =>
+	booleanWithDetailsSummaryListItem({
+		id: 'health-and-safety',
+		text: 'Potential safety risks',
+		value: lpaQuestionnaireData.doesSiteHaveHealthAndSafetyIssues,
+		valueDetails: lpaQuestionnaireData.healthAndSafetyDetails,
+		defaultText: 'No answer provided',
+		link: `${currentRoute}/safety-risks/change/lpa`,
+		editable: userHasUpdateCase,
+		addCyAttribute: true,
+		classes: 'lpa-health-and-safety'
+	});
