+++ appeals/web/src/server/lib/mappers/lpa-questionnaire/submappers/map-plans-drawings.js
@@ -1,0 +1,11 @@
+import { documentInstruction } from '../utils.js';
+
+/** @type {import("../lpa-questionnaire.mapper.js").SubMapper} */
+export const mapPlansDrawings = ({ lpaQuestionnaireData, session }) =>
+	documentInstruction({
+		id: 'plans-drawings',
+		text: `Plans, drawings and list of plans`,
+		folderInfo: lpaQuestionnaireData.documents.plansDrawings,
+		lpaQuestionnaireData,
+		session
+	});
