+++ appeals/web/src/server/lib/mappers/lpa-questionnaire/submappers/map-press-advert.js
@@ -1,0 +1,12 @@
+import { documentInstruction } from '../utils.js';
+
+/** @type {import("../lpa-questionnaire.mapper.js").SubMapper} */
+export const mapPressAdvert = ({ lpaQuestionnaireData, session }) =>
+	documentInstruction({
+		id: 'press-advert',
+		text: 'Press advert notification',
+		folderInfo: lpaQuestionnaireData.documents.whoNotifiedPressAdvert,
+		cypressDataName: 'press-advert-notification',
+		lpaQuestionnaireData,
+		session
+	});
