+++ appeals/web/src/server/lib/mappers/lpa-questionnaire/submappers/map-site-within-green-belt.js
@@ -1,0 +1,13 @@
+import { booleanSummaryListItem } from '#lib/mappers/components/boolean.js';
+
+/** @type {import("../lpa-questionnaire.mapper.js").SubMapper} */
+export const mapSiteWithinGreenBelt = ({ lpaQuestionnaireData, currentRoute, userHasUpdateCase }) =>
+	booleanSummaryListItem({
+		id: 'site-within-green-belt',
+		text: 'Green belt',
+		value: lpaQuestionnaireData.isGreenBelt,
+		defaultText: '',
+		addCyAttribute: true,
+		link: `${currentRoute}/green-belt/change/lpa`,
+		editable: userHasUpdateCase
+	});
